unit-test: Add the selftest to detect subsytem
To establish the testing, it first needs to define the corresponding
macro and include the unit_test.h at the bottom of the target file we
want to test. Then, in the unit_test.h, we need to include the test suit
file located in the tests directory. For the test suit file, we will use
the specific io function provided by the printut.h. After finishing all
the settings, we execute the script unit_test.sh in the tests directory.

Here is the example:

ucsan/tests$ ./unit_test.sh
a - src/core.o
a - src/unify.o
a - lib/per_cpu.o
a - tests/test_watchpoint.o

 ------------------------------ unit test start ------------------------------

 ============================== detect ==============================

 [1655284647995497] INFO: src/core.c:89:test_insert_find_watchpoint(): ptr=0x7fff3efd87ac size=4 type=1 watchpoint=0x55e81bc3e100
 [1655284647995563] INFO: src/core.c:105:test_insert_find_watchpoint(): watchpoint=0x55e81bc3e100 old=8007fff3efd87ac new=9007fff3efd87ac
 [TEST SUCCESS] test_insert_find_watchpoint()
 [1655284647995743] INFO: src/core.c:65:test_setup_report(): ptr=0x7fff3efd87c4 size=4 type=1
 [1655284647995775] INFO: src/core.c:42:test_setup_report_thread(): ptr=0x7fff3efd87c4 size=4 type=1, watchpoint=(idx=24,0x55e81bc3e100)
 [1655284647995897] INFO: src/core.c:21:unify_report(): ptr=0x7fff3efd87c4, size=4, type=1, ip=12345, old=8007fff3efd87c4, new=9007fff3efd87c4, changed=1
 [TEST SUCCESS] test_setup_report()

 ============================== detect ==============================

 ------------------------------ unit test end --------------------------------

Also, simplify some code in src/core.c.

Signed-off-by: Chih-En Lin <shiyn.lin@gmail.com>