detect: Change condition of data race
Currently, the data race condition is to check the watchpoint is
modified. And send the watchpoint state to the unify subsystem. But the
unify subsystem cannot use this information to get the old value.

So, here we change the logic of how the data race detects. We compare
the value before and after the delay to determine whether the data race
occurs. And for this new logic, we also don't have to use the operation
type in the watchpoint state to see the data race happens (write/write,
write/read pairs).

Signed-off-by: Chih-En Lin <shiyn.lin@gmail.com>